Removed dummy panels, code cleaning.
