+++ piplmesh/frontend/templates/home.html
@@ -20,10 +20,7 @@
 {% block content %}
     <div id="wall">
     </div>
-    <div id="spinner" style="display: block;">
-        <img src="../static/piplmesh/images/ajax-loader.gif" />
-    </div>
-    <div id="panels" style="visibility: hidden;">
+    <div id="panels">
         <div class="panels_column">
             {% for panel in user.get_panels %}
                 {% render_panel panel %}
