+++ piplmesh/panels/dummy/panel.py
@@ -1,5 +1,6 @@
 import random
 
+from django.conf import settings
 from django.contrib.webdesign import lorem_ipsum
 from django.utils.translation import ugettext_lazy as _
 
@@ -14,4 +15,35 @@ class DummyPanel(panels.BasePanel):
         })
         return context
 
-panels.panels_pool.register(DummyPanel)
+class DummyPanel2(panels.BasePanel):
+    def get_context(self, context):
+        context.update({
+            'header': _("Dummy panel 2"),
+            'content': '\n\n'.join(lorem_ipsum.paragraphs(random.randint(1, 1))),
+            'id': 'dummy2',
+        })
+        return context
+
+class DummyPanel3(panels.BasePanel):
+    def get_context(self, context):
+        context.update({
+            'header': _("Dummy panel 3"),
+            'content': '\n\n'.join(lorem_ipsum.paragraphs(random.randint(1, 1))),
+            'id': 'dummy3',
+        })
+        return context
+
+class DummyPanel4(panels.BasePanel):
+    def get_context(self, context):
+        context.update({
+            'header': _("Dummy panel 4"),
+            'content': '\n\n'.join(lorem_ipsum.paragraphs(random.randint(1, 1))),
+            'id': 'dummy4',
+        })
+        return context
+
+if settings.DEBUG:
+    panels.panels_pool.register(DummyPanel)
+    panels.panels_pool.register(DummyPanel2)
+    panels.panels_pool.register(DummyPanel3)
+    panels.panels_pool.register(DummyPanel4)
