--- piplmesh/panels/dummy2/templates/panel/dummy/panel.html
@@ -1,5 +1,0 @@
-{% extends "panel/panel.html" %}
-
-{% block content %}
-    {{ content|linebreaks }}
-{% endblock %}
