--- piplmesh/panels/dummy4/panel.py
@@ -1,17 +1,0 @@
-import random
-
-from django.contrib.webdesign import lorem_ipsum
-from django.utils.translation import ugettext_lazy as _
-
-from piplmesh import panels
-
-class DummyPanel4(panels.BasePanel):
-    def get_context(self, context):
-        context.update({
-            'header': _("Dummy panel 4"),
-            'content': '\n\n'.join(lorem_ipsum.paragraphs(random.randint(1, 1))),
-            'id': 'dummy4',
-        })
-        return context
-
-panels.panels_pool.register(DummyPanel4)
